refactor: extract Binance public archive acquisition seam
